# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMMAS API",
        "description": "Sistem Informasi Manajemen Magang Siswa (SIMMAS). REST API dengan autentikasi Sanctum. Token didapat dari endpoint `POST /api/login`. Semua endpoint selain `login` dan `public/*` memerlukan header `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server backend SIMMAS (Laravel)"
        }
    ],
    "paths": {
...
        "/api/admin/dudi": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Daftar mitra DUDI",
                "description": "Daftar semua tempat magang beserta jumlah penempatan. Dapat difilter dengan pencarian.",
                "operationId": "102bde94f765490663161fd6e3938881",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama perusahaan atau bidang usaha.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar DUDI.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TempatMagang"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tambah mitra DUDI",
                "operationId": "e561849298c1b743cd913d11661b4961",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_perusahaan",
                                    "bidang_usaha",
                                    "nama_pic",
                                    "kontak_pic",
                                    "kuota",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama_perusahaan": {
                                        "type": "string",
                                        "example": "PT Teknologi Nusantara"
                                    },
                                    "bidang_usaha": {
                                        "type": "string",
                                        "example": "Teknologi Informasi"
                                    },
                                    "nama_pic": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "kontak_pic": {
                                        "type": "string",
                                        "example": "0812-3456-7890"
                                    },
                                    "kuota": {
                                        "type": "integer",
                                        "example": 10,
                                        "minimum": 1
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Industri No. 10, Malang"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DUDI berhasil ditambahkan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TempatMagang"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dudi/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Perbarui data mitra DUDI",
                "operationId": "15cac479a24b595a87fdb32ebdcdaaf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_perusahaan",
                                    "bidang_usaha",
                                    "nama_pic",
                                    "kontak_pic",
                                    "kuota",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama_perusahaan": {
                                        "type": "string"
                                    },
                                    "bidang_usaha": {
                                        "type": "string"
                                    },
                                    "nama_pic": {
                                        "type": "string"
                                    },
                                    "kontak_pic": {
                                        "type": "string"
                                    },
                                    "kuota": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "alamat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "DUDI berhasil diperbarui.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TempatMagang"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus mitra DUDI",
                "description": "Penghapusan ditolak apabila masih ada siswa yang sedang aktif magang di DUDI tersebut.",
                "operationId": "8a5ab3c0f73355c97bb48dcb97744836",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DUDI berhasil dihapus."
                    },
                    "422": {
                        "description": "Gagal: masih ada siswa aktif magang."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/dudi/{id}/verifikasi": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Toggle status verifikasi DUDI",
                "operationId": "e59932bcf6cb8398e04869d61d17318d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status verifikasi berhasil diubah.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TempatMagang"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TempatMagang": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nama_perusahaan": {
                        "type": "string"
                    },
                    "bidang_usaha": {
                        "type": "string"
                    },
                    "nama_pic": {
                        "type": "string"
                    },
                    "kontak_pic": {
                        "type": "string"
                    },
                    "kuota": {
                        "type": "integer"
                    },
                    "sisa_kuota": {
                        "type": "integer"
                    },
                    "alamat": {
                        "type": "string"
                    },
                    "status_verifikasi": {
                        "type": "string",
                        "enum": [
                            "terverifikasi",
                            "belum_diverifikasi"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer <token> (didapat dari endpoint POST /api/login).",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}